requirements.txt
discord.py
psycopg2
pytz
requests
langid
cryptography